% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpb_crs.R
\name{base_crs}
\alias{base_crs}
\title{CRS values}
\usage{
base_crs()
}
\value{
a number matching the EPSG code of the selected CRS
}
\description{
\code{base_crs()} returns the coordinate reference system (CRS)
to be used within Wikiplantbaser, i.e. WGS84 / UTM zone 32
(epsg: 32632), which covers the Italian territory.
}
\details{
\code{point_crs()} returns the CRS of point occurrences as
extracted from wikiplantbase.
}
\examples{
base_crs <- base_crs()
}
