% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpb_getByTaxon.R
\name{wpb_getByTaxon}
\alias{wpb_getByTaxon}
\title{Easy access to Wikiplantbase}
\usage{
wpb_getByTaxon(taxon = c("Crocus etruscus Parl."))
}
\arguments{
\item{taxon}{A vector of valid scientific names of vascular plant
according to the online Checklist of the Ittalian flora
(https://dryades.units.it/floritaly). Partial names (e.g. "Crocus il")
will be retrieved as all names starting with the partial value.}
}
\value{
A dataframe
}
\description{
Get georeferenced records for taxa of the Italian vascular flora
passed as character vector. It is not meant for massive downloads:
you should be able to obtain several thousands records in seconds,
but for tens or hundreds of thousands of records please contact
Wikiplantbase coordinators.
}
\examples{
wpb_getByTaxon(c("Crocus etruscus Parl.","Crocus il"))
}
