% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpb_postRequest.R
\name{wpb_postRequest}
\alias{wpb_postRequest}
\title{Send a PST request to Wikiplantbase API}
\usage{
wpb_postRequest(url, body, gc_prec)
}
\arguments{
\item{url}{the URL to send the request to}

\item{body}{the data, as list(taxon=your_taxon_list) in JSON format}

\item{gc_prec}{the precision of geographic coordinates (number of decimals of lat/lon values)}
}
\value{
a spatial dataframe
}
\description{
This function is not to be exposed to users; it is used by
wpb_getByXxxxx functions to retrieve data from Wikiplantbase
}
\examples{
url <- "http://bot.biologia.unipi.it/wpbold/php/wpb_getByTaxon.php"
body <- rjson::toJSON(list(taxon=c("Crocus etruscus")))
gc_prec <- 2
w <- wpb_postRequest(url = url, body = body, gc_prec = gc_prec)
}
