% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpb_plot.R
\name{wpb_plot}
\alias{wpb_plot}
\title{Plot occurrence points}
\usage{
wpb_plot(xx, ...)
}
\arguments{
\item{xx}{a spatial dataframe such as that produced by \code{wpb_getByTaxon}}

\item{...}{other parameters passed on}
}
\value{
a ggplot2 object
}
\description{
Plots occurrence points on a map of Italy
}
\examples{
ce_plot <- wpb_plot(wpb_crocEtrOcc)
}
